EDIT: api 파일 코드 수정
- 잘못된 함수 위치 변경
- bearer 사용 유/무 수정
