+++ frontend/src/api/product.ts
@@ -16,15 +16,6 @@ export const postProducts = async (formData: PostObjectType, token: string) => {
   return res;
 };
 
-export const getMembers = async (token: string | null) => {
-  const res = await axiosInstanceWithBearer.get(`/api/members`, {
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
-  });
-  return res;
-};
-
 export const getProductDetail = async (
   productsId: string | undefined,
   token: string | null,
@@ -34,18 +25,6 @@ export const getProductDetail = async (
   }
 
   const res = await axiosInstanceWithBearer.get(`/api/products/${productsId}`, {
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
-  });
-  return res;
-};
-
-export const getSeller = async (
-  token: string | null,
-  curRoomId: string | undefined,
-) => {
-  const res = await axiosInstanceWithBearer.get(`chat/room/${curRoomId}`, {
     headers: {
       Authorization: `Bearer ${token}`,
     },
