+++ frontend/src/pages/ChatRoom/index.tsx
@@ -11,7 +11,7 @@ import ChatInputBar from '../../components/ChatInputBar';
 import NavBarTitle from '../../components/NavBarTitle';
 import { ACCESS_TOKEN } from '../../constants/login';
 import useAsync from '../../hooks/useAsync';
-import { getSeller } from '../../api/product';
+import { getSeller } from '../../api/member';
 
 interface ChatHistoryProps {
   type: string;
